requirements.txt
fastapi
uvicorn
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
Pillow
transformers 
python-multipart
timm 
einops
accelerate>=0.26.0